refactor(go style guide) : refactor code base in alignment to go style guide
