+++ handlers/sdkwrapper.go
@@ -4,11 +4,35 @@ import (
 	"encoding/json"
 	"fmt"
 	"net/http"
+	"time"
 
 	"github.com/aws/aws-sdk-go-v2/service/iam"
 	"github.com/rajasoun/aws-hub/handlers/api"
+	"github.com/rajasoun/aws-hub/service/aws"
 	"github.com/rajasoun/aws-hub/service/cache"
 )
+
+type AWSHandler struct {
+	cache    cache.Cache
+	multiple bool
+	aws      aws.AWS
+}
+
+func NewDefaultAWSHandler(multiple bool) *AWSHandler {
+	defaultCacheDuration := 30
+	cacheHandler := &cache.Memory{Expiration: time.Duration(defaultCacheDuration)}
+	cacheHandler.Connect()
+	return NewAWSHandler(cacheHandler, multiple)
+}
+
+func NewAWSHandler(cacheHandler cache.Cache, multiple bool) *AWSHandler {
+	awsHandler := AWSHandler{
+		cache:    cacheHandler,
+		multiple: multiple,
+		aws:      aws.AWS{},
+	}
+	return &awsHandler
+}
 
 type AWSWrapper struct {
 	request  *http.Request
