+++ handlers/sdkwrapper_test.go
@@ -7,8 +7,37 @@ import (
 
 	"github.com/aws/aws-sdk-go-v2/service/iam"
 	"github.com/rajasoun/aws-hub/test"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 )
+
+func TestNewDefaultAWSHandler(t *testing.T) {
+	assert := assert.New(t)
+	t.Parallel()
+	tests := []struct {
+		name          string
+		multiple      bool
+		wantCacheType string
+	}{
+		{
+			name:          "Check New Default AWS Handler with no profile",
+			multiple:      false,
+			wantCacheType: "InMemoryCache",
+		},
+		{
+			name:          "Check New Default AWS Handler with multiple profile",
+			multiple:      true,
+			wantCacheType: "InMemoryCache",
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got := NewDefaultAWSHandler(tt.multiple)
+			assert.Equal(got.cache.Type(), tt.wantCacheType,
+				"NewDefaultAWSHandler() = %v, want %v", got, tt.wantCacheType)
+		})
+	}
+}
 
 // To mock AWS operations.
 type MockAwsAPI struct {
